+++ src/COSMOS_SDK_COMMIT
@@ -1,1 +1,1 @@
-666c345ad23ddda9523cc5cd1b71187d91c26f34
+2e9e5d6eea24d6c11eddc9c002c66e89ae036187
