+++ src/prost/google.protobuf.rs
@@ -44,14 +44,9 @@ pub struct FileDescriptorProto {
     #[prost(message, optional, tag = "9")]
     pub source_code_info: ::core::option::Option<SourceCodeInfo>,
     /// The syntax of the proto file.
-    /// The supported values are "proto2", "proto3", and "editions".
-    ///
-    /// If `edition` is present, this value must be "editions".
+    /// The supported values are "proto2" and "proto3".
     #[prost(string, optional, tag = "12")]
     pub syntax: ::core::option::Option<::prost::alloc::string::String>,
-    /// The edition of the proto file, which is an opaque string.
-    #[prost(string, optional, tag = "13")]
-    pub edition: ::core::option::Option<::prost::alloc::string::String>,
 }
 /// Describes a message type.
 #[allow(clippy::derive_partial_eq_without_eq)]
@@ -114,92 +109,6 @@ pub struct ExtensionRangeOptions {
     /// The parser stores options it doesn't recognize here. See above.
     #[prost(message, repeated, tag = "999")]
     pub uninterpreted_option: ::prost::alloc::vec::Vec<UninterpretedOption>,
-    /// go/protobuf-stripping-extension-declarations
-    /// Like Metadata, but we use a repeated field to hold all extension
-    /// declarations. This should avoid the size increases of transforming a large
-    /// extension range into small ranges in generated binaries.
-    #[prost(message, repeated, tag = "2")]
-    pub declaration: ::prost::alloc::vec::Vec<extension_range_options::Declaration>,
-    /// The verification state of the range.
-    /// TODO(b/278783756): flip the default to DECLARATION once all empty ranges
-    /// are marked as UNVERIFIED.
-    #[prost(
-        enumeration = "extension_range_options::VerificationState",
-        optional,
-        tag = "3",
-        default = "Unverified"
-    )]
-    pub verification: ::core::option::Option<i32>,
-}
-/// Nested message and enum types in `ExtensionRangeOptions`.
-pub mod extension_range_options {
-    #[allow(clippy::derive_partial_eq_without_eq)]
-    #[derive(Clone, PartialEq, ::prost::Message)]
-    pub struct Declaration {
-        /// The extension number declared within the extension range.
-        #[prost(int32, optional, tag = "1")]
-        pub number: ::core::option::Option<i32>,
-        /// The fully-qualified name of the extension field. There must be a leading
-        /// dot in front of the full name.
-        #[prost(string, optional, tag = "2")]
-        pub full_name: ::core::option::Option<::prost::alloc::string::String>,
-        /// The fully-qualified type name of the extension field. Unlike
-        /// Metadata.type, Declaration.type must have a leading dot for messages
-        /// and enums.
-        #[prost(string, optional, tag = "3")]
-        pub r#type: ::core::option::Option<::prost::alloc::string::String>,
-        /// Deprecated. Please use "repeated".
-        #[deprecated]
-        #[prost(bool, optional, tag = "4")]
-        pub is_repeated: ::core::option::Option<bool>,
-        /// If true, indicates that the number is reserved in the extension range,
-        /// and any extension field with the number will fail to compile. Set this
-        /// when a declared extension field is deleted.
-        #[prost(bool, optional, tag = "5")]
-        pub reserved: ::core::option::Option<bool>,
-        /// If true, indicates that the extension must be defined as repeated.
-        /// Otherwise the extension must be defined as optional.
-        #[prost(bool, optional, tag = "6")]
-        pub repeated: ::core::option::Option<bool>,
-    }
-    /// The verification state of the extension range.
-    #[derive(
-        Clone,
-        Copy,
-        Debug,
-        PartialEq,
-        Eq,
-        Hash,
-        PartialOrd,
-        Ord,
-        ::prost::Enumeration
-    )]
-    #[repr(i32)]
-    pub enum VerificationState {
-        /// All the extensions of the range must be declared.
-        Declaration = 0,
-        Unverified = 1,
-    }
-    impl VerificationState {
-        /// String value of the enum field names used in the ProtoBuf definition.
-        ///
-        /// The values are not transformed in any way and thus are considered stable
-        /// (if the ProtoBuf definition does not change) and safe for programmatic use.
-        pub fn as_str_name(&self) -> &'static str {
-            match self {
-                VerificationState::Declaration => "DECLARATION",
-                VerificationState::Unverified => "UNVERIFIED",
-            }
-        }
-        /// Creates an enum from field names used in the ProtoBuf definition.
-        pub fn from_str_name(value: &str) -> ::core::option::Option<Self> {
-            match value {
-                "DECLARATION" => Some(Self::Declaration),
-                "UNVERIFIED" => Some(Self::Unverified),
-                _ => None,
-            }
-        }
-    }
 }
 /// Describes a field within a message.
 #[allow(clippy::derive_partial_eq_without_eq)]
@@ -699,10 +608,6 @@ pub struct MessageOptions {
     /// this is a formalization for deprecating messages.
     #[prost(bool, optional, tag = "3", default = "false")]
     pub deprecated: ::core::option::Option<bool>,
-    /// NOTE: Do not set the option in .proto files. Always use the maps syntax
-    /// instead. The option should only be implicitly set by the proto compiler
-    /// parser.
-    ///
     /// Whether the message is an automatically generated map entry type for the
     /// maps field.
     ///
@@ -720,21 +625,12 @@ pub struct MessageOptions {
     /// use a native map in the target language to hold the keys and values.
     /// The reflection APIs in such implementations still need to work as
     /// if the field is a repeated message field.
+    ///
+    /// NOTE: Do not set the option in .proto files. Always use the maps syntax
+    /// instead. The option should only be implicitly set by the proto compiler
+    /// parser.
     #[prost(bool, optional, tag = "7")]
     pub map_entry: ::core::option::Option<bool>,
-    /// Enable the legacy handling of JSON field name conflicts.  This lowercases
-    /// and strips underscored from the fields before comparison in proto3 only.
-    /// The new behavior takes `json_name` into account and applies to proto2 as
-    /// well.
-    ///
-    /// This should only be used as a temporary measure against broken builds due
-    /// to the change in behavior for JSON field name conflicts.
-    ///
-    /// TODO(b/261750190) This is legacy behavior we plan to remove once downstream
-    /// teams have had time to migrate.
-    #[deprecated]
-    #[prost(bool, optional, tag = "11")]
-    pub deprecated_legacy_json_field_conflicts: ::core::option::Option<bool>,
     /// The parser stores options it doesn't recognize here. See above.
     #[prost(message, repeated, tag = "999")]
     pub uninterpreted_option: ::prost::alloc::vec::Vec<UninterpretedOption>,
@@ -744,10 +640,8 @@ pub struct MessageOptions {
 pub struct FieldOptions {
     /// The ctype option instructs the C++ code generator to use a different
     /// representation of the field than it normally would.  See the specific
-    /// options below.  This option is only implemented to support use of
-    /// \[ctype=CORD\] and \[ctype=STRING\] (the default) on non-repeated fields of
-    /// type "bytes" in the open source release -- sorry, we'll try to include
-    /// other types in a future version!
+    /// options below.  This option is not yet implemented in the open source
+    /// release -- sorry, we'll try to include it in a future version!
     #[prost(
         enumeration = "field_options::CType",
         optional,
@@ -797,6 +691,7 @@ pub struct FieldOptions {
     /// call from multiple threads concurrently, while non-const methods continue
     /// to require exclusive access.
     ///
+    ///
     /// Note that implementations may choose not to check required fields within
     /// a lazy sub-message.  That is, calling IsInitialized() on the outer message
     /// may return true even if the inner message has missing required fields.
@@ -808,8 +703,11 @@ pub struct FieldOptions {
     /// check its required fields, regardless of whether or not the message has
     /// been parsed.
     ///
-    /// As of May 2022, lazy verifies the contents of the byte stream during
-    /// parsing.  An invalid byte stream will cause the overall parsing to fail.
+    /// As of 2021, lazy does no correctness checks on the byte stream during
+    /// parsing.  This may lead to crashes if and when an invalid byte stream is
+    /// finally parsed upon access.
+    ///
+    /// TODO(b/211906113):  Enable validation on lazy fields.
     #[prost(bool, optional, tag = "5", default = "false")]
     pub lazy: ::core::option::Option<bool>,
     /// unverified_lazy does no correctness checks on the byte stream. This should
@@ -826,22 +724,6 @@ pub struct FieldOptions {
     /// For Google-internal migration only. Do not use.
     #[prost(bool, optional, tag = "10", default = "false")]
     pub weak: ::core::option::Option<bool>,
-    /// Indicate that the field value should not be printed out when using debug
-    /// formats, e.g. when the field contains sensitive credentials.
-    #[prost(bool, optional, tag = "16", default = "false")]
-    pub debug_redact: ::core::option::Option<bool>,
-    #[prost(enumeration = "field_options::OptionRetention", optional, tag = "17")]
-    pub retention: ::core::option::Option<i32>,
-    #[deprecated]
-    #[prost(enumeration = "field_options::OptionTargetType", optional, tag = "18")]
-    pub target: ::core::option::Option<i32>,
-    #[prost(
-        enumeration = "field_options::OptionTargetType",
-        repeated,
-        packed = "false",
-        tag = "19"
-    )]
-    pub targets: ::prost::alloc::vec::Vec<i32>,
     /// The parser stores options it doesn't recognize here. See above.
     #[prost(message, repeated, tag = "999")]
     pub uninterpreted_option: ::prost::alloc::vec::Vec<UninterpretedOption>,
@@ -863,12 +745,6 @@ pub mod field_options {
     pub enum CType {
         /// Default mode.
         String = 0,
-        /// The option \[ctype=CORD\] may be applied to a non-repeated field of type
-        /// "bytes". It indicates that in C++, the data should be stored in a Cord
-        /// instead of a string.  For very large strings, this may reduce memory
-        /// fragmentation. It may also allow better performance when parsing from a
-        /// Cord, or when parsing with aliasing enabled, as the parsed Cord may then
-        /// alias the original buffer.
         Cord = 1,
         StringPiece = 2,
     }
@@ -936,114 +812,6 @@ pub mod field_options {
             }
         }
     }
-    /// If set to RETENTION_SOURCE, the option will be omitted from the binary.
-    /// Note: as of January 2023, support for this is in progress and does not yet
-    /// have an effect (b/264593489).
-    #[derive(
-        Clone,
-        Copy,
-        Debug,
-        PartialEq,
-        Eq,
-        Hash,
-        PartialOrd,
-        Ord,
-        ::prost::Enumeration
-    )]
-    #[repr(i32)]
-    pub enum OptionRetention {
-        RetentionUnknown = 0,
-        RetentionRuntime = 1,
-        RetentionSource = 2,
-    }
-    impl OptionRetention {
-        /// String value of the enum field names used in the ProtoBuf definition.
-        ///
-        /// The values are not transformed in any way and thus are considered stable
-        /// (if the ProtoBuf definition does not change) and safe for programmatic use.
-        pub fn as_str_name(&self) -> &'static str {
-            match self {
-                OptionRetention::RetentionUnknown => "RETENTION_UNKNOWN",
-                OptionRetention::RetentionRuntime => "RETENTION_RUNTIME",
-                OptionRetention::RetentionSource => "RETENTION_SOURCE",
-            }
-        }
-        /// Creates an enum from field names used in the ProtoBuf definition.
-        pub fn from_str_name(value: &str) -> ::core::option::Option<Self> {
-            match value {
-                "RETENTION_UNKNOWN" => Some(Self::RetentionUnknown),
-                "RETENTION_RUNTIME" => Some(Self::RetentionRuntime),
-                "RETENTION_SOURCE" => Some(Self::RetentionSource),
-                _ => None,
-            }
-        }
-    }
-    /// This indicates the types of entities that the field may apply to when used
-    /// as an option. If it is unset, then the field may be freely used as an
-    /// option on any kind of entity. Note: as of January 2023, support for this is
-    /// in progress and does not yet have an effect (b/264593489).
-    #[derive(
-        Clone,
-        Copy,
-        Debug,
-        PartialEq,
-        Eq,
-        Hash,
-        PartialOrd,
-        Ord,
-        ::prost::Enumeration
-    )]
-    #[repr(i32)]
-    pub enum OptionTargetType {
-        TargetTypeUnknown = 0,
-        TargetTypeFile = 1,
-        TargetTypeExtensionRange = 2,
-        TargetTypeMessage = 3,
-        TargetTypeField = 4,
-        TargetTypeOneof = 5,
-        TargetTypeEnum = 6,
-        TargetTypeEnumEntry = 7,
-        TargetTypeService = 8,
-        TargetTypeMethod = 9,
-    }
-    impl OptionTargetType {
-        /// String value of the enum field names used in the ProtoBuf definition.
-        ///
-        /// The values are not transformed in any way and thus are considered stable
-        /// (if the ProtoBuf definition does not change) and safe for programmatic use.
-        pub fn as_str_name(&self) -> &'static str {
-            match self {
-                OptionTargetType::TargetTypeUnknown => "TARGET_TYPE_UNKNOWN",
-                OptionTargetType::TargetTypeFile => "TARGET_TYPE_FILE",
-                OptionTargetType::TargetTypeExtensionRange => {
-                    "TARGET_TYPE_EXTENSION_RANGE"
-                }
-                OptionTargetType::TargetTypeMessage => "TARGET_TYPE_MESSAGE",
-                OptionTargetType::TargetTypeField => "TARGET_TYPE_FIELD",
-                OptionTargetType::TargetTypeOneof => "TARGET_TYPE_ONEOF",
-                OptionTargetType::TargetTypeEnum => "TARGET_TYPE_ENUM",
-                OptionTargetType::TargetTypeEnumEntry => "TARGET_TYPE_ENUM_ENTRY",
-                OptionTargetType::TargetTypeService => "TARGET_TYPE_SERVICE",
-                OptionTargetType::TargetTypeMethod => "TARGET_TYPE_METHOD",
-            }
-        }
-        /// Creates an enum from field names used in the ProtoBuf definition.
-        pub fn from_str_name(value: &str) -> ::core::option::Option<Self> {
-            match value {
-                "TARGET_TYPE_UNKNOWN" => Some(Self::TargetTypeUnknown),
-                "TARGET_TYPE_FILE" => Some(Self::TargetTypeFile),
-                "TARGET_TYPE_EXTENSION_RANGE" => Some(Self::TargetTypeExtensionRange),
-                "TARGET_TYPE_MESSAGE" => Some(Self::TargetTypeMessage),
-                "TARGET_TYPE_FIELD" => Some(Self::TargetTypeField),
-                "TARGET_TYPE_ONEOF" => Some(Self::TargetTypeOneof),
-                "TARGET_TYPE_ENUM" => Some(Self::TargetTypeEnum),
-                "TARGET_TYPE_ENUM_ENTRY" => Some(Self::TargetTypeEnumEntry),
-                "TARGET_TYPE_SERVICE" => Some(Self::TargetTypeService),
-                "TARGET_TYPE_METHOD" => Some(Self::TargetTypeMethod),
-                _ => None,
-            }
-        }
-    }
 }
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
@@ -1065,15 +833,6 @@ pub struct EnumOptions {
     /// is a formalization for deprecating enums.
     #[prost(bool, optional, tag = "3", default = "false")]
     pub deprecated: ::core::option::Option<bool>,
-    /// Enable the legacy handling of JSON field name conflicts.  This lowercases
-    /// and strips underscored from the fields before comparison in proto3 only.
-    /// The new behavior takes `json_name` into account and applies to proto2 as
-    /// well.
-    /// TODO(b/261750190) Remove this legacy behavior once downstream teams have
-    /// had time to migrate.
-    #[deprecated]
-    #[prost(bool, optional, tag = "6")]
-    pub deprecated_legacy_json_field_conflicts: ::core::option::Option<bool>,
     /// The parser stores options it doesn't recognize here. See above.
     #[prost(message, repeated, tag = "999")]
     pub uninterpreted_option: ::prost::alloc::vec::Vec<UninterpretedOption>,
@@ -1202,8 +961,8 @@ pub mod uninterpreted_option {
     /// The name of the uninterpreted option.  Each string represents a segment in
     /// a dot-separated name.  is_extension is true iff a segment represents an
     /// extension (denoted with parentheses in options specs in .proto files).
-    /// E.g.,{ ["foo", false], ["bar.baz", true], ["moo", false] } represents
-    /// "foo.(bar.baz).moo".
+    /// E.g.,{ ["foo", false], ["bar.baz", true], ["qux", false] } represents
+    /// "foo.(bar.baz).qux".
     #[allow(clippy::derive_partial_eq_without_eq)]
     #[derive(Clone, PartialEq, ::prost::Message)]
     pub struct NamePart {
@@ -1328,13 +1087,13 @@ pub mod source_code_info {
         ///    // Comment attached to baz.
         ///    // Another line attached to baz.
         ///
-        ///    // Comment attached to moo.
+        ///    // Comment attached to qux.
         ///    //
-        ///    // Another line attached to moo.
-        ///    optional double moo = 4;
+        ///    // Another line attached to qux.
+        ///    optional double qux = 4;
         ///
         ///    // Detached comment for corge. This is not leading or trailing comments
-        ///    // to moo or corge because there are blank lines separating it from
+        ///    // to qux or corge because there are blank lines separating it from
         ///    // both.
         ///
         ///    // Detached comment for corge paragraph 2.
@@ -1386,59 +1145,10 @@ pub mod generated_code_info {
         #[prost(int32, optional, tag = "3")]
         pub begin: ::core::option::Option<i32>,
         /// Identifies the ending offset in bytes in the generated code that
-        /// relates to the identified object. The end offset should be one past
+        /// relates to the identified offset. The end offset should be one past
         /// the last relevant byte (so the length of the text = end - begin).
         #[prost(int32, optional, tag = "4")]
         pub end: ::core::option::Option<i32>,
-        #[prost(enumeration = "annotation::Semantic", optional, tag = "5")]
-        pub semantic: ::core::option::Option<i32>,
-    }
-    /// Nested message and enum types in `Annotation`.
-    pub mod annotation {
-        /// Represents the identified object's effect on the element in the original
-        /// .proto file.
-        #[derive(
-            Clone,
-            Copy,
-            Debug,
-            PartialEq,
-            Eq,
-            Hash,
-            PartialOrd,
-            Ord,
-            ::prost::Enumeration
-        )]
-        #[repr(i32)]
-        pub enum Semantic {
-            /// There is no effect or the effect is indescribable.
-            None = 0,
-            /// The element is set or otherwise mutated.
-            Set = 1,
-            /// An alias to the element is returned.
-            Alias = 2,
-        }
-        impl Semantic {
-            /// String value of the enum field names used in the ProtoBuf definition.
-            ///
-            /// The values are not transformed in any way and thus are considered stable
-            /// (if the ProtoBuf definition does not change) and safe for programmatic use.
-            pub fn as_str_name(&self) -> &'static str {
-                match self {
-                    Semantic::None => "NONE",
-                    Semantic::Set => "SET",
-                    Semantic::Alias => "ALIAS",
-                }
-            }
-            /// Creates an enum from field names used in the ProtoBuf definition.
-            pub fn from_str_name(value: &str) -> ::core::option::Option<Self> {
-                match value {
-                    "NONE" => Some(Self::None),
-                    "SET" => Some(Self::Set),
-                    "ALIAS" => Some(Self::Alias),
-                    _ => None,
-                }
-            }
-        }
     }
 }
 /// `Any` contains an arbitrary serialized protocol buffer message along with a
@@ -1465,12 +1175,8 @@ pub mod generated_code_info {
 ///      if (any.is(Foo.class)) {
 ///        foo = any.unpack(Foo.class);
 ///      }
-///      // or ...
-///      if (any.isSameTypeAs(Foo.getDefaultInstance())) {
-///        foo = any.unpack(Foo.getDefaultInstance());
-///      }
 ///
-///   Example 3: Pack and unpack a message in Python.
+/// Example 3: Pack and unpack a message in Python.
 ///
 ///      foo = Foo(...)
 ///      any = Any()
@@ -1480,7 +1186,7 @@ pub mod generated_code_info {
 ///        any.Unpack(foo)
 ///        ...
 ///
-///   Example 4: Pack and unpack a message in Go
+/// Example 4: Pack and unpack a message in Go
 ///
 ///       foo := &pb.Foo{...}
 ///       any, err := anypb.New(foo)
@@ -1499,8 +1205,9 @@ pub mod generated_code_info {
 /// in the type URL, for example "foo.bar.com/x/y.z" will yield type
 /// name "y.z".
 ///
+///
 /// JSON
-/// ====
+///
 /// The JSON representation of an `Any` value uses the regular
 /// representation of the deserialized, embedded message, with an
 /// additional field `@type` which contains the type URL. Example:
@@ -1616,6 +1323,7 @@ pub struct Any {
 ///      Timestamp timestamp = Timestamp.newBuilder().setSeconds(millis / 1000)
 ///          .setNanos((int) ((millis % 1000) * 1000000)).build();
 ///
+///
 /// Example 5: Compute Timestamp from Java `Instant.now()`.
 ///
 ///      Instant now = Instant.now();
@@ -1623,6 +1331,7 @@ pub struct Any {
 ///      Timestamp timestamp =
 ///          Timestamp.newBuilder().setSeconds(now.getEpochSecond())
 ///              .setNanos(now.getNano()).build();
+///
 ///
 /// Example 6: Compute Timestamp from current time in Python.
 ///
@@ -1653,8 +1362,9 @@ pub struct Any {
 /// \[`strftime`\](<https://docs.python.org/2/library/time.html#time.strftime>) with
 /// the time format spec '%Y-%m-%dT%H:%M:%S.%fZ'. Likewise, in Java, one can use
 /// the Joda Time's \[`ISODateTimeFormat.dateTime()`\](
-/// <http://joda-time.sourceforge.net/apidocs/org/joda/time/format/ISODateTimeFormat.html#dateTime(>)
+/// <http://www.joda.org/joda-time/apidocs/org/joda/time/format/ISODateTimeFormat.html#dateTime%2D%2D>
 /// ) to obtain a formatter capable of generating timestamps in this format.
+///
 ///
 #[cfg_attr(feature = "serde", derive(::serde::Serialize, ::serde::Deserialize))]
 #[allow(clippy::derive_partial_eq_without_eq)]
@@ -1730,6 +1440,7 @@ pub struct Timestamp {
 /// encoded in JSON format as "3s", while 3 seconds and 1 nanosecond should
 /// be expressed in JSON format as "3.000000001s", and 3 seconds and 1
 /// microsecond should be expressed in JSON format as "3.000001s".
+///
 ///
 #[cfg_attr(feature = "serde", derive(::serde::Serialize, ::serde::Deserialize))]
 #[derive(Eq)]
