+++ src/prost/ibc.applications.transfer.v1.rs
@@ -82,9 +82,9 @@ pub struct MsgTransferResponse {
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct MsgUpdateParams {
-    /// authority is the address that controls the module (defaults to x/gov unless overwritten).
+    /// signer address (it may be the the address that controls the module, which defaults to x/gov unless overwritten).
     #[prost(string, tag = "1")]
-    pub authority: ::prost::alloc::string::String,
+    pub signer: ::prost::alloc::string::String,
     /// params defines the transfer parameters to update.
     ///
     /// NOTE: All parameters must be supplied.
