+++ src/prost/ibc.core.client.v1.rs
@@ -289,9 +289,9 @@ pub struct MsgSubmitMisbehaviourResponse {}
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct MsgUpdateParams {
-    /// authority is the address of the governance account.
+    /// signer address (it may be the the address that controls the module, which defaults to x/gov unless overwritten).
     #[prost(string, tag = "1")]
-    pub authority: ::prost::alloc::string::String,
+    pub signer: ::prost::alloc::string::String,
     /// params defines the client parameters to update.
     ///
     /// NOTE: All parameters must be supplied.
