+++ src/prost/interchain_security.ccv.v1.rs
@@ -17,13 +17,6 @@ pub struct ValidatorSetChangePacketData {
     #[prost(string, repeated, tag = "3")]
     pub slash_acks: ::prost::alloc::vec::Vec<::prost::alloc::string::String>,
 }
-/// List of ccv.ValidatorSetChangePacketData.
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct ValidatorSetChangePackets {
-    #[prost(message, repeated, tag = "1")]
-    pub list: ::prost::alloc::vec::Vec<ValidatorSetChangePacketData>,
-}
 /// This packet is sent from the consumer chain to the provider chain
 /// to notify that a VSC packet reached maturity on the consumer chain.
 #[allow(clippy::derive_partial_eq_without_eq)]
@@ -51,15 +44,7 @@ pub struct SlashPacketData {
     )]
     pub infraction: i32,
 }
-/// MaturedUnbondingOps defines a list of ids corresponding to ids of matured
-/// unbonding operations.
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct MaturedUnbondingOps {
-    #[prost(uint64, repeated, tag = "1")]
-    pub ids: ::prost::alloc::vec::Vec<u64>,
-}
-/// ConsumerPacketData contains a consumer packet data, type tag, and index for storage.
+/// ConsumerPacketData contains a consumer packet data and a type tag
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct ConsumerPacketData {
@@ -79,13 +64,14 @@ pub mod consumer_packet_data {
         VscMaturedPacketData(super::VscMaturedPacketData),
     }
 }
-/// ConsumerPacketDataList is a list of consumer packet data packets.
-/// NOTE: It is only used for exporting / importing state in InitGenesis and ExportGenesis.
+/// Note this type is used during IBC handshake methods for both the consumer and provider
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
-pub struct ConsumerPacketDataList {
-    #[prost(message, repeated, tag = "1")]
-    pub list: ::prost::alloc::vec::Vec<ConsumerPacketData>,
+pub struct HandshakeMetadata {
+    #[prost(string, tag = "1")]
+    pub provider_fee_pool_addr: ::prost::alloc::string::String,
+    #[prost(string, tag = "2")]
+    pub version: ::prost::alloc::string::String,
 }
 /// ConsumerPacketData contains a consumer packet data and a type tag
 /// that is compatible with ICS v1 and v2 over the wire. It is not used for internal storage.
@@ -162,7 +148,7 @@ impl ConsumerPacketDataType {
     }
 }
 /// InfractionType indicates the infraction type a validator commited.
-/// NOTE: ccv.InfractionType to maintain compatibility between ICS versions
+/// Note ccv.InfractionType to maintain compatibility between ICS versions
 /// using different versions of the cosmos-sdk and ibc-go modules.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
 #[repr(i32)]
@@ -195,4 +181,183 @@ impl InfractionType {
             _ => None,
         }
     }
+}
+/// Params defines the parameters for CCV consumer module.
+///
+/// Note this type is referenced in both the consumer and provider CCV modules,
+/// and persisted on the provider, see MakeConsumerGenesis and SetConsumerGenesis.
+///
+/// TODO: Rename to ConsumerParams. See <https://github.com/cosmos/interchain-security/issues/1206>
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct Params {
+    /// TODO: Remove enabled flag and find a better way to setup integration tests
+    /// See: <https://github.com/cosmos/interchain-security/issues/339>
+    #[prost(bool, tag = "1")]
+    pub enabled: bool,
+    /// /////////////////////
+    /// Distribution Params
+    /// Number of blocks between ibc-token-transfers from the consumer chain to
+    /// the provider chain. Note that at this transmission event a fraction of
+    /// the accumulated tokens are divided and sent consumer redistribution
+    /// address.
+    #[prost(int64, tag = "2")]
+    pub blocks_per_distribution_transmission: i64,
+    /// Channel, and provider-chain receiving address to send distribution token
+    /// transfers over. These parameters is auto-set during the consumer <->
+    /// provider handshake procedure.
+    #[prost(string, tag = "3")]
+    pub distribution_transmission_channel: ::prost::alloc::string::String,
+    #[prost(string, tag = "4")]
+    pub provider_fee_pool_addr_str: ::prost::alloc::string::String,
+    /// Sent CCV related IBC packets will timeout after this duration
+    #[prost(message, optional, tag = "5")]
+    pub ccv_timeout_period: ::core::option::Option<
+        super::super::super::google::protobuf::Duration,
+    >,
+    /// Sent transfer related IBC packets will timeout after this duration
+    #[prost(message, optional, tag = "6")]
+    pub transfer_timeout_period: ::core::option::Option<
+        super::super::super::google::protobuf::Duration,
+    >,
+    /// The fraction of tokens allocated to the consumer redistribution address
+    /// during distribution events. The fraction is a string representing a
+    /// decimal number. For example "0.75" would represent 75%.
+    #[prost(string, tag = "7")]
+    pub consumer_redistribution_fraction: ::prost::alloc::string::String,
+    /// The number of historical info entries to persist in store.
+    /// This param is a part of the cosmos sdk staking module. In the case of
+    /// a ccv enabled consumer chain, the ccv module acts as the staking module.
+    #[prost(int64, tag = "8")]
+    pub historical_entries: i64,
+    /// Unbonding period for the consumer,
+    /// which should be smaller than that of the provider in general.
+    #[prost(message, optional, tag = "9")]
+    pub unbonding_period: ::core::option::Option<
+        super::super::super::google::protobuf::Duration,
+    >,
+    /// The threshold for the percentage of validators at the bottom of the set who
+    /// can opt out of running the consumer chain without being punished. For
+    /// example, a value of 0.05 means that the validators in the bottom 5% of the
+    /// set can opt out
+    #[prost(string, tag = "10")]
+    pub soft_opt_out_threshold: ::prost::alloc::string::String,
+    /// Reward denoms. These are the denominations which are allowed to be sent to
+    /// the provider as rewards.
+    #[prost(string, repeated, tag = "11")]
+    pub reward_denoms: ::prost::alloc::vec::Vec<::prost::alloc::string::String>,
+    /// Provider-originated reward denoms. These are denoms coming from the
+    /// provider which are allowed to be used as rewards. e.g. "uatom"
+    #[prost(string, repeated, tag = "12")]
+    pub provider_reward_denoms: ::prost::alloc::vec::Vec<::prost::alloc::string::String>,
+}
+/// GenesisState defines the CCV consumer chain genesis state.
+///
+/// Note this type is referenced in both the consumer and provider CCV modules,
+/// and persisted on the provider, see MakeConsumerGenesis and SetConsumerGenesis.
+///
+/// TODO: Rename to ConsumerGenesisState. See <https://github.com/cosmos/interchain-security/issues/1206>
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct GenesisState {
+    #[prost(message, optional, tag = "1")]
+    pub params: ::core::option::Option<Params>,
+    /// empty for a new chain, filled in on restart.
+    #[prost(string, tag = "2")]
+    pub provider_client_id: ::prost::alloc::string::String,
+    /// empty for a new chain, filled in on restart.
+    #[prost(string, tag = "3")]
+    pub provider_channel_id: ::prost::alloc::string::String,
+    /// true for new chain GenesisState, false for chain restart.
+    #[prost(bool, tag = "4")]
+    pub new_chain: bool,
+    /// ProviderClientState filled in on new chain, nil on restart.
+    #[prost(message, optional, tag = "5")]
+    pub provider_client_state: ::core::option::Option<
+        super::super::super::ibc::lightclients::tendermint::v1::ClientState,
+    >,
+    /// ProviderConsensusState filled in on new chain, nil on restart.
+    #[prost(message, optional, tag = "6")]
+    pub provider_consensus_state: ::core::option::Option<
+        super::super::super::ibc::lightclients::tendermint::v1::ConsensusState,
+    >,
+    /// MaturingPackets nil on new chain, filled in on restart.
+    #[prost(message, repeated, tag = "7")]
+    pub maturing_packets: ::prost::alloc::vec::Vec<MaturingVscPacket>,
+    /// InitialValset filled in on new chain and on restart.
+    #[prost(message, repeated, tag = "8")]
+    pub initial_val_set: ::prost::alloc::vec::Vec<
+        ::tendermint_proto::abci::ValidatorUpdate,
+    >,
+    /// HeightToValsetUpdateId nil on new chain, filled in on restart.
+    #[prost(message, repeated, tag = "9")]
+    pub height_to_valset_update_id: ::prost::alloc::vec::Vec<HeightToValsetUpdateId>,
+    /// OutstandingDowntimes nil on new chain, filled  in on restart.
+    #[prost(message, repeated, tag = "10")]
+    pub outstanding_downtime_slashing: ::prost::alloc::vec::Vec<OutstandingDowntime>,
+    /// PendingConsumerPackets nil on new chain, filled in on restart.
+    #[prost(message, optional, tag = "11")]
+    pub pending_consumer_packets: ::core::option::Option<ConsumerPacketDataList>,
+    /// LastTransmissionBlockHeight nil on new chain, filled in on restart.
+    #[prost(message, optional, tag = "12")]
+    pub last_transmission_block_height: ::core::option::Option<
+        LastTransmissionBlockHeight,
+    >,
+    /// flag indicating whether the consumer CCV module starts in
+    #[prost(bool, tag = "13")]
+    pub pre_ccv: bool,
+}
+/// HeightValsetUpdateID represents a mapping internal to the consumer CCV module
+/// AND used in shared consumer genesis state, which links a block height to each recv valset update id.
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct HeightToValsetUpdateId {
+    #[prost(uint64, tag = "1")]
+    pub height: u64,
+    #[prost(uint64, tag = "2")]
+    pub valset_update_id: u64,
+}
+/// OutstandingDowntime defines the type used internally to the consumer CCV module,
+/// AND used in shared consumer genesis state, in order to not send multiple slashing
+/// requests for the same downtime infraction.
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct OutstandingDowntime {
+    #[prost(string, tag = "1")]
+    pub validator_consensus_address: ::prost::alloc::string::String,
+}
+/// LastTransmissionBlockHeight is the last time validator holding
+/// pools were transmitted to the provider chain. This type is used internally
+/// to the consumer CCV module AND used in shared consumer genesis state.
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct LastTransmissionBlockHeight {
+    #[prost(int64, tag = "1")]
+    pub height: i64,
+}
+/// MaturingVSCPacket represents a vsc packet that is maturing internal to the
+/// consumer CCV module, where the consumer has not yet relayed a VSCMatured packet
+/// back to the provider.  This type is used internally to the consumer CCV module
+/// AND used in shared consumer genesis state.
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct MaturingVscPacket {
+    #[prost(uint64, tag = "1")]
+    pub vsc_id: u64,
+    #[prost(message, optional, tag = "2")]
+    pub maturity_time: ::core::option::Option<
+        super::super::super::google::protobuf::Timestamp,
+    >,
+}
+/// ConsumerPacketDataList is a list of consumer packet data packets.
+///
+/// Note this type is is used internally to the consumer CCV module
+/// for exporting / importing state in InitGenesis and ExportGenesis,
+/// AND included in the consumer genesis type (reffed by provider and consumer modules),
+/// hence this is a shared type.
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct ConsumerPacketDataList {
+    #[prost(message, repeated, tag = "1")]
+    pub list: ::prost::alloc::vec::Vec<ConsumerPacketData>,
 }
